push up

+++ src/main/java/mrriegel/limelib/helper/ColorHelper.java
@@ -45,7 +45,7 @@ public class ColorHelper {
 		// return new Color(color).getAlpha();
 	}
 
-	public static int getRainbow(long frequence) {
+	public static int getRainbow(int frequence) {
 		if (frequence == 0l)
 			frequence = 1;
 		return Color.getHSBColor(((System.currentTimeMillis() / frequence) % 360l) / 360f, 1, 1).getRGB();
